// apps/frontend/src/lib/api/generated/models/ServiceList.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Lightweight serializer for service lists
 */
export type ServiceList = {
    readonly id: number;
    title: string;
    slug: string;
    readonly icon_url: string;
    readonly options_count: string;
};